green: Add Restaurants component
